[NO-ISSUE] Style based on guidelines (#1)
* feat(aside-container): initial styling of aside container

Styles the aside container with the main content (logo + title). Also, sets up the footer of it with
some styling to enhance readability.

* feat(*): set ratio of aside to main content

Will cap the min size of the aside at 400px but will group half as much as the main container.

* feat(*): add style guide and update the main container

Set styles based on assignment defaults and prettify while leveraging css custom properties.